Refactor sidebar and weapon chip into modules
